+++ src/ResourceManagement/Application/WarehouseManagement.ResourceManagement.Application/UseCases/DeleteResource/DeleteResourceHandler.cs
@@ -6,6 +6,8 @@ using WarehouseManagement.Core.Abstractions;
 using WarehouseManagement.Core.Abstractions.Messages;
 using WarehouseManagement.Core.Enums;
 using WarehouseManagement.Core.Extensions;
+using WarehouseManagement.IncomeProcessing.Contracts;
+using WarehouseManagement.OutcomeProcessing.Contracts;
 using WarehouseManagement.ResourceManagement.Domain.Entities;
 using WarehouseManagement.SharedKernel;
 using WarehouseManagement.SharedKernel.ValueObjects.Ids;
@@ -14,6 +16,10 @@ namespace WarehouseManagement.ResourceManagement.Application.UseCases.DeleteReso
 
 public class DeleteResourceHandler : ICommandHandler<Guid, DeleteResourceCommand>
 {
+    private readonly IIncomeProcessingContract _incomeProcessingContract;
+    
+    private readonly IOutcomeProcessingContract _outcomeProcessingContract;
+    
     private readonly IRepository<Resource,ResourceId> _resourceRepository;
     
     private readonly IUnitOfWork _unitOfWork;
@@ -26,12 +32,16 @@ public class DeleteResourceHandler : ICommandHandler<Guid, DeleteResourceCommand
         IRepository<Resource, ResourceId> resourceRepository,
         [FromKeyedServices(Modules.ResourceManagement)] IUnitOfWork unitOfWork, 
         IValidator<DeleteResourceCommand> validator,
-        ILogger<DeleteResourceHandler> logger)
+        ILogger<DeleteResourceHandler> logger,
+        IIncomeProcessingContract incomeProcessingContract,
+        IOutcomeProcessingContract outcomeProcessingContract)
     {
         _resourceRepository = resourceRepository;
         _unitOfWork = unitOfWork;
         _validator = validator;
         _logger = logger;
+        _incomeProcessingContract = incomeProcessingContract;
+        _outcomeProcessingContract = outcomeProcessingContract;
     }
 
     public async Task<Result<Guid, ErrorList>> Handle(DeleteResourceCommand command, CancellationToken cancellationToken = default)
@@ -49,6 +59,16 @@ public class DeleteResourceHandler : ICommandHandler<Guid, DeleteResourceCommand
             return resourceResult.Error.ToErrorList();
         
         var resource =  resourceResult.Value;
+        
+        var checkResourceIdNotUsedInAnyIncomeResourceResult = await _incomeProcessingContract.CheckResourceIdNotUsedInAnyIncomeResource(resourceId);
+        
+        if (checkResourceIdNotUsedInAnyIncomeResourceResult.IsFailure) 
+            return checkResourceIdNotUsedInAnyIncomeResourceResult.Error.ToErrorList();
+        
+        var checkResourceIdNotUsedInAnyOutcomeResourceResult = await _outcomeProcessingContract.CheckResourceIdNotUsedInAnyOutcomeResource(resourceId);
+        
+        if (checkResourceIdNotUsedInAnyOutcomeResourceResult.IsFailure) 
+            return checkResourceIdNotUsedInAnyOutcomeResourceResult.Error.ToErrorList();
         
         _resourceRepository.Delete(resource);
