+++ src/UnitManagement/Application/WarehouseManagement.UnitManagement.Application/UseCases/DeleteUnit/DeleteUnitHandler.cs
@@ -6,6 +6,8 @@ using WarehouseManagement.Core.Abstractions;
 using WarehouseManagement.Core.Abstractions.Messages;
 using WarehouseManagement.Core.Enums;
 using WarehouseManagement.Core.Extensions;
+using WarehouseManagement.IncomeProcessing.Contracts;
+using WarehouseManagement.OutcomeProcessing.Contracts;
 using WarehouseManagement.SharedKernel;
 using WarehouseManagement.SharedKernel.ValueObjects.Ids;
 using WarehouseManagement.UnitManagement.Domain.Entities;
@@ -14,6 +16,10 @@ namespace WarehouseManagement.UnitManagement.Application.UseCases.DeleteUnit;
 
 public class DeleteUnitHandler : ICommandHandler<Guid,  DeleteUnitCommand>
 {
+    private readonly IIncomeProcessingContract _incomeProcessingContract;
+    
+    private readonly IOutcomeProcessingContract _outcomeProcessingContract;
+    
     private readonly IRepository<Unit,UnitId> _unitRepository;
     
     private readonly IUnitOfWork _unitOfWork;
@@ -26,12 +32,16 @@ public class DeleteUnitHandler : ICommandHandler<Guid,  DeleteUnitCommand>
         IRepository<Unit, UnitId> unitRepository, 
         [FromKeyedServices(Modules.UnitManagement)] IUnitOfWork unitOfWork,
         IValidator<DeleteUnitCommand> validator,
-        ILogger<DeleteUnitHandler> logger)
+        ILogger<DeleteUnitHandler> logger,
+        IIncomeProcessingContract incomeProcessingContract,
+        IOutcomeProcessingContract outcomeProcessingContract)
     {
         _unitRepository = unitRepository;
         _unitOfWork = unitOfWork;
         _validator = validator;
         _logger = logger;
+        _incomeProcessingContract = incomeProcessingContract;
+        _outcomeProcessingContract = outcomeProcessingContract;
     }
 
     public async Task<Result<Guid, ErrorList>> Handle(DeleteUnitCommand command, CancellationToken cancellationToken = default)
@@ -49,6 +59,16 @@ public class DeleteUnitHandler : ICommandHandler<Guid,  DeleteUnitCommand>
             return unitResult.Error.ToErrorList();
         
         var unit =  unitResult.Value;
+        
+        var checkUnitIdNotUsedInAnyIncomeResourceResult = await _incomeProcessingContract.CheckUnitIdNotUsedInAnyIncomeResource(unitId);
+        
+        if (checkUnitIdNotUsedInAnyIncomeResourceResult.IsFailure) 
+            return checkUnitIdNotUsedInAnyIncomeResourceResult.Error.ToErrorList();
+
+        var checkUnitIdNotUsedInAnyOutcomeResourceResult = await _outcomeProcessingContract.CheckUnitIdNotUsedInAnyOutcomeResource(unitId);
+        
+        if (checkUnitIdNotUsedInAnyOutcomeResourceResult.IsFailure) 
+            return checkUnitIdNotUsedInAnyOutcomeResourceResult.Error.ToErrorList();
         
         _unitRepository.Delete(unit);
